refactor tests back to handler files
