+++ .github/workflows/unittests.yml
@@ -1,6 +1,9 @@
 name: Contract Tests
 
 on:
+  push:
+    branches:
+      - main
   pull_request:
   workflow_dispatch:
