+++ contracts/dca/src/handlers/mod.rs
@@ -7,6 +7,7 @@ pub mod deposit;
 pub mod disburse_escrow;
 pub mod disburse_funds;
 pub mod execute_trigger;
+pub mod get_config;
 pub mod get_custom_swap_fees;
 pub mod get_dca_plus_performance;
 pub mod get_disburse_escrow_tasks;
@@ -17,6 +18,8 @@ pub mod get_time_trigger_ids;
 pub mod get_vault;
 pub mod get_vaults;
 pub mod get_vaults_by_address;
+pub mod instantiate;
+pub mod migrate;
 pub mod remove_custom_swap_fee;
 pub mod update_config;
 pub mod update_swap_adjustments_handler;
