--- contracts/dca/src/tests/get_vaults_tests.rs
@@ -1,189 +1,0 @@
-use crate::handlers::get_vaults::get_vaults_handler;
-use crate::tests::helpers::{instantiate_contract, setup_new_vault};
-use crate::tests::mocks::ADMIN;
-use crate::types::vault::Vault;
-use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
-use cosmwasm_std::Uint128;
-
-#[test]
-fn with_limit_too_large_should_fail() {
-    let mut deps = mock_dependencies();
-
-    instantiate_contract(deps.as_mut(), mock_env(), mock_info(ADMIN, &[]));
-
-    let err = get_vaults_handler(deps.as_ref(), None, Some(1001)).unwrap_err();
-
-    assert_eq!(
-        err.to_string(),
-        "Generic error: limit cannot be greater than 1000."
-    );
-}
-
-#[test]
-fn with_no_vaults_should_return_all_vaults() {
-    let mut deps = mock_dependencies();
-
-    instantiate_contract(deps.as_mut(), mock_env(), mock_info(ADMIN, &[]));
-
-    let vaults = get_vaults_handler(deps.as_ref(), None, None)
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.len(), 0);
-}
-
-#[test]
-fn with_multiple_vaults_should_return_all_vaults() {
-    let mut deps = mock_dependencies();
-    let env = mock_env();
-
-    instantiate_contract(deps.as_mut(), env.clone(), mock_info(ADMIN, &[]));
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(1),
-            ..Vault::default()
-        },
-    );
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(2),
-            ..Vault::default()
-        },
-    );
-
-    let vaults = get_vaults_handler(deps.as_ref(), None, None)
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.len(), 2);
-}
-
-#[test]
-fn with_one_vault_should_return_proper_vault_data() {
-    let mut deps = mock_dependencies();
-    let env = mock_env();
-
-    instantiate_contract(deps.as_mut(), env.clone(), mock_info(ADMIN, &[]));
-
-    let vault = setup_new_vault(deps.as_mut(), env.clone(), Vault::default());
-
-    let vaults = get_vaults_handler(deps.as_ref(), None, None)
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.first().unwrap(), &vault);
-}
-
-#[test]
-fn with_limit_should_return_limited_vaults() {
-    let mut deps = mock_dependencies();
-    let env = mock_env();
-
-    instantiate_contract(deps.as_mut(), env.clone(), mock_info(ADMIN, &[]));
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(1),
-            ..Vault::default()
-        },
-    );
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(2),
-            ..Vault::default()
-        },
-    );
-
-    let vaults = get_vaults_handler(deps.as_ref(), None, Some(1))
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.len(), 1);
-    assert_eq!(vaults[0].id, Uint128::new(1));
-}
-
-#[test]
-fn with_start_after_should_return_vaults_after_start_after() {
-    let mut deps = mock_dependencies();
-    let env = mock_env();
-
-    instantiate_contract(deps.as_mut(), env.clone(), mock_info(ADMIN, &[]));
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(1),
-            ..Vault::default()
-        },
-    );
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(2),
-            ..Vault::default()
-        },
-    );
-
-    let vaults = get_vaults_handler(deps.as_ref(), Some(1), None)
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.len(), 1);
-    assert_eq!(vaults[0].id, Uint128::new(2));
-}
-
-#[test]
-fn with_limit_and_start_after_should_return_limited_vaults_after_start_after() {
-    let mut deps = mock_dependencies();
-    let env = mock_env();
-
-    instantiate_contract(deps.as_mut(), env.clone(), mock_info(ADMIN, &[]));
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(1),
-            ..Vault::default()
-        },
-    );
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(2),
-            ..Vault::default()
-        },
-    );
-
-    setup_new_vault(
-        deps.as_mut(),
-        env.clone(),
-        Vault {
-            id: Uint128::new(3),
-            ..Vault::default()
-        },
-    );
-
-    let vaults = get_vaults_handler(deps.as_ref(), Some(1), Some(1))
-        .unwrap()
-        .vaults;
-
-    assert_eq!(vaults.len(), 1);
-    assert_eq!(vaults[0].id, Uint128::new(2));
-}
