+++ contracts/dca/src/tests/mod.rs
@@ -3,60 +3,3 @@ pub mod mocks;
 
 #[cfg(test)]
 pub mod helpers;
-
-#[cfg(test)]
-pub mod instantiate_tests;
-
-#[cfg(test)]
-pub mod create_vault_tests;
-
-#[cfg(test)]
-pub mod cancel_vault_tests;
-
-#[cfg(test)]
-pub mod execute_trigger_tests;
-
-#[cfg(test)]
-pub mod get_vaults_by_address_tests;
-
-#[cfg(test)]
-pub mod deposit_tests;
-
-#[cfg(test)]
-pub mod update_config_tests;
-
-#[cfg(test)]
-pub mod disburse_funds_tests;
-
-#[cfg(test)]
-pub mod get_events_by_resource_id_tests;
-
-#[cfg(test)]
-pub mod create_custom_swap_fee_tests;
-
-#[cfg(test)]
-pub mod remove_custom_fee_tests;
-
-#[cfg(test)]
-pub mod get_vaults_tests;
-
-#[cfg(test)]
-pub mod get_config_tests;
-
-#[cfg(test)]
-pub mod disburse_escrow_tests;
-
-#[cfg(test)]
-pub mod update_swap_adjustments_tests;
-
-#[cfg(test)]
-pub mod create_pair_tests;
-
-#[cfg(test)]
-mod get_pairs_tests;
-
-#[cfg(test)]
-pub mod z_provide_liquidity_tests;
-
-#[cfg(test)]
-pub mod z_delegate_tests;
