--- contracts/dca/src/tests/scenario_tests.rs
@@ -1,8 +1,0 @@
-#[test]
-fn vault_with_price_trigger_executes_until_empty() {
-    // create vault with fin trigger
-
-    // execute fin trigger
-
-    // advance time & execute price trggers until vault is zero
-}
